requirements.txt
pydantic>=2.0.0
asyncio
typing-extensions
langgraph
langchain-google-genai
langchain
streamlit
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
pydantic==2.5.0
pydantic-settings==2.1.0
jinja2==3.1.2
python-multipart==0.0.6
schedule==1.2.0

# Oracle Cloud Infrastructure
oci

# Data processing and visualization
pandas==2.1.4
numpy==1.24.3
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# PDF generation
pdfkit==1.0.0
weasyprint==60.2

# MCP (Model Context Protocol)
mcp==0.3.0

# Additional utilities
python-dotenv==1.0.0
aiofiles==23.2.1
# Remove any pdfkit or wkhtmltopdf references
matplotlib>=3.7.0
reportlab>=4.0.0
Pillow>=10.0.0
# Remove: seaborn (causing issues)

